feat: move package checks from var-naming to package-naming
