+++ test/var_naming_test.go
@@ -42,38 +42,6 @@ func TestVarNaming(t *testing.T) {
 	testRule(t, "var_naming_upper_case_const_true", &rule.VarNamingRule{}, &lint.RuleConfig{
 		Arguments: lint.Arguments{[]any{}, []any{}, []any{map[string]any{"upper-case-const": true}}},
 	})
-
-	testRule(t, "var_naming_skip_package_name_checks_false", &rule.VarNamingRule{}, &lint.RuleConfig{})
-	testRule(t, "var_naming_skip_package_name_checks_true", &rule.VarNamingRule{}, &lint.RuleConfig{
-		Arguments: lint.Arguments{[]any{}, []any{}, []any{map[string]any{"skipPackageNameChecks": true}}},
-	})
-	testRule(t, "var_naming_skip_package_name_checks_true", &rule.VarNamingRule{}, &lint.RuleConfig{
-		Arguments: lint.Arguments{[]any{}, []any{}, []any{map[string]any{"skip-package-name-checks": true}}},
-	})
-	testRule(t, "var_naming_meaningless_package_name", &rule.VarNamingRule{}, &lint.RuleConfig{})
-	testRule(t, "var_naming_meaningless_package_name", &rule.VarNamingRule{}, &lint.RuleConfig{
-		Arguments: lint.Arguments{
-			[]any{},
-			[]any{},
-			[]any{map[string]any{"skip-package-name-checks": false}},
-		},
-	})
-	testRule(t, "var_naming_bad_package_name", &rule.VarNamingRule{}, &lint.RuleConfig{
-		Arguments: lint.Arguments{
-			[]any{},
-			[]any{},
-			[]any{map[string]any{"skip-package-name-checks": false, "extra-bad-package-names": []any{"helpers"}}},
-		},
-	})
-	testRule(t, "var_naming_top_level_pkg", &rule.VarNamingRule{}, &lint.RuleConfig{})
-	testRule(t, "var_naming_std_lib_conflict", &rule.VarNamingRule{}, &lint.RuleConfig{})
-	testRule(t, "var_naming_std_lib_conflict_skip", &rule.VarNamingRule{}, &lint.RuleConfig{
-		Arguments: lint.Arguments{
-			[]any{},
-			[]any{},
-			[]any{map[string]any{"skip-package-name-collision-with-go-std": true}},
-		},
-	})
 }
 
 func BenchmarkUpperCaseConstTrue(b *testing.B) {
