+++ testdata/package_naming_a_pkg_with_caps.go
