+++ testdata/package_naming_bad_package_name.go
