+++ testdata/package_naming_meaningless_package_name.go
