+++ testdata/package_naming_skip_package_name_checks_true.go
