+++ testdata/package_naming_std_lib_conflict.go
