+++ testdata/package_naming_top_level_pkg.go
